+++ js/templates/game-1/game-1.js
@@ -4,6 +4,8 @@ import {getCurrentStateGame1, getCurrentLevel} from '../../current-state.js';
 import Game1View from './game-1-view.js';
 import {intro} from '../intro/intro.js';
 import {stats} from '../stats/stats.js';
+import answersPoint from '../stats/stats-answers.js';
+import Timer from '../../timer.js';
 
 const game1 = (game) => {
   const viewGame1 = new Game1View(game);
@@ -18,6 +20,8 @@ const game1 = (game) => {
   };
 
   viewGame1.nextView = () => {
+    game.timer.stop();
+    game.timer = new Timer(30);
     showWindow(game2(game));
   };
 
@@ -26,6 +30,7 @@ const game1 = (game) => {
   };
 
   viewGame1.statsView = () => {
+    answersPoint(game);
     showWindow(stats(game));
   };
