+++ js/templates/game-2/game-2.js
@@ -4,6 +4,8 @@ import ViewGame2 from './game-2-view.js';
 import {intro} from '../intro/intro.js';
 import {game3} from '../game-3/game-3.js';
 import {stats} from '../stats/stats.js';
+import answersPoint from '../stats/stats-answers.js';
+import Timer from '../../timer.js';
 
 const game2 = (game) => {
   const viewGame2 = new ViewGame2(game);
@@ -18,6 +20,8 @@ const game2 = (game) => {
   };
 
   viewGame2.nextView = () => {
+    game.timer.stop();
+    game.timer = new Timer(30);
     showWindow(game3(game));
   };
 
@@ -26,6 +30,7 @@ const game2 = (game) => {
   };
 
   viewGame2.statsView = () => {
+    answersPoint(game);
     showWindow(stats(game));
   };
