+++ js/templates/game-3/game-3.js
@@ -4,6 +4,8 @@ import {game1} from '../game-1/game-1.js';
 import {getCurrentStateGame3, getCurrentLevel} from '../../current-state.js';
 import ViewGame3 from './game-3-view.js';
 import {intro} from '../intro/intro.js';
+import answersPoint from '../stats/stats-answers.js';
+import Timer from '../../timer.js';
 
 const game3 = (game) => {
   const viewGame3 = new ViewGame3(game);
@@ -18,6 +20,8 @@ const game3 = (game) => {
   };
 
   viewGame3.nextView = () => {
+    game.timer.stop();
+    game.timer = new Timer(30);
     showWindow(game1(game));
   };
 
@@ -26,6 +30,7 @@ const game3 = (game) => {
   };
 
   viewGame3.statsView = () => {
+    answersPoint(game);
     showWindow(stats(game));
   };
